requirements.txt
socketio
python-socketio[asyncio_client]
LZString
python-dotenv
